SONARPY-786 RSPEC-3827: Extract detection of never defined variable in a separate rule (RSPEC-5953)
