+++ its/ruling/src/test/resources/expected/python-S5953.json
@@ -1,0 +1,134 @@
+{
+'project:S1716.py':[
+4,
+],
+'project:buildbot-0.8.6p1/buildbot/test/unit/test_db_sourcestamps.py':[
+120,
+],
+'project:mypy-0.782/test-data/stdlib-samples/3.2/test/test_tempfile.py':[
+262,
+568,
+],
+'project:numpy-1.16.4/numpy/distutils/ccompiler.py':[
+214,
+696,
+],
+'project:numpy-1.16.4/numpy/distutils/npy_pkg_config.py':[
+249,
+249,
+],
+'project:numpy-1.16.4/numpy/distutils/unixccompiler.py':[
+15,
+133,
+],
+'project:numpy-1.16.4/numpy/linalg/lapack_lite/clapack_scrub.py':[
+306,
+],
+'project:numpy-1.16.4/numpy/matlib.py':[
+49,
+171,
+352,
+],
+'project:tensorflow/__init__.py':[
+39,
+40,
+],
+'project:tensorflow/python/autograph/converters/control_flow_test.py':[
+680,
+],
+'project:tensorflow/python/autograph/pyct/cfg_test.py':[
+1224,
+1279,
+],
+'project:tensorflow/python/autograph/pyct/static_analysis/activity_test.py':[
+199,
+],
+'project:tensorflow/python/autograph/pyct/static_analysis/liveness_test.py':[
+510,
+],
+'project:tensorflow/python/autograph/pyct/static_analysis/reaching_definitions_test.py':[
+133,
+373,
+],
+'project:twisted-12.1.0/doc/core/howto/listings/pb/trap_client.py':[
+21,
+21,
+],
+'project:twisted-12.1.0/doc/core/howto/tutorial/listings/finger/finger19a_changes.py':[
+2,
+9,
+11,
+11,
+17,
+27,
+28,
+28,
+],
+'project:twisted-12.1.0/doc/core/howto/tutorial/listings/finger/finger19b_changes.py':[
+7,
+9,
+9,
+],
+'project:twisted-12.1.0/doc/core/howto/tutorial/listings/finger/finger19c_changes.py':[
+8,
+10,
+10,
+],
+'project:twisted-12.1.0/doc/core/howto/tutorial/listings/finger/finger_config.py':[
+6,
+7,
+8,
+8,
+12,
+14,
+20,
+25,
+35,
+35,
+],
+'project:twisted-12.1.0/doc/historic/2003/pycon/deferex/deferex-bad-adding.py':[
+8,
+],
+'project:twisted-12.1.0/doc/historic/2003/pycon/deferex/deferex-forwarding.py':[
+3,
+],
+'project:twisted-12.1.0/doc/historic/2003/pycon/deferex/deferex-listing0.py':[
+4,
+13,
+],
+'project:twisted-12.1.0/doc/historic/2003/pycon/deferex/deferex-listing2.py':[
+8,
+],
+'project:twisted-12.1.0/doc/web/examples/users.rpy.py':[
+24,
+],
+'project:twisted-12.1.0/twisted/conch/scripts/cftp.py':[
+431,
+],
+'project:twisted-12.1.0/twisted/conch/scripts/ckeygen.py':[
+75,
+],
+'project:twisted-12.1.0/twisted/conch/scripts/tkconch.py':[
+376,
+],
+'project:twisted-12.1.0/twisted/python/compat.py':[
+22,
+],
+'project:twisted-12.1.0/twisted/runner/inetdtap.py':[
+58,
+59,
+62,
+152,
+],
+'project:twisted-12.1.0/twisted/test/proto_helpers.py':[
+412,
+414,
+],
+'project:twisted-12.1.0/twisted/test/test_amp.py':[
+2985,
+2989,
+],
+'project:twisted-12.1.0/twisted/test/test_paths.py':[
+527,
+],
+}
