+++ python-checks/src/main/resources/org/sonar/l10n/py/rules/python/S5953.json
@@ -1,0 +1,12 @@
+{
+  "title": "Variables, classes and functions should be either defined or imported",
+  "type": "BUG",
+  "status": "ready",
+  "tags": [
+    
+  ],
+  "defaultSeverity": "Blocker",
+  "ruleSpecification": "RSPEC-5953",
+  "sqKey": "S5953",
+  "scope": "All"
+}
