+++ python-checks/src/test/java/org/sonar/python/checks/UndeclaredNameUsageCheckTest.java
@@ -19,7 +19,6 @@
  */
 package org.sonar.python.checks;
 
-import java.util.Arrays;
 import org.junit.Test;
 import org.sonar.python.checks.utils.PythonCheckVerifier;
 
@@ -28,30 +27,6 @@ public class UndeclaredNameUsageCheckTest {
   @Test
   public void test() {
     PythonCheckVerifier.verify("src/test/resources/checks/undeclaredNameUsage.py", new UndeclaredNameUsageCheck());
-  }
-
-  @Test
-  public void test_wildcard_import() {
-    PythonCheckVerifier.verify(
-      Arrays.asList("src/test/resources/checks/undeclaredNameUsageWithWildcardImport.py","src/test/resources/checks/undeclaredNameUsageImported.py" ),
-      new UndeclaredNameUsageCheck());
-  }
-
-  @Test
-  public void test_wildcard_import_all_property() {
-    PythonCheckVerifier.verifyNoIssue(
-      Arrays.asList("src/test/resources/checks/undeclaredNameUsageImportWithAll.py", "src/test/resources/checks/undeclaredNameUsageAll/__init__.py"),
-      new UndeclaredNameUsageCheck());
-  }
-
-  @Test
-  public void test_unresolved_wildcard_import() {
-    PythonCheckVerifier.verifyNoIssue("src/test/resources/checks/undeclaredNameUsageWithUnresolvedWildcardImport.py", new UndeclaredNameUsageCheck());
-  }
-
-  @Test
-  public void test_dynamic_globals() {
-    PythonCheckVerifier.verifyNoIssue("src/test/resources/checks/undeclaredNameUsageWithGlobals.py", new UndeclaredNameUsageCheck());
   }
 
 }
