+++ python-checks/src/test/java/org/sonar/python/checks/UnresolvedSymbolsCheckTest.java
@@ -1,0 +1,57 @@
+/*
+ * SonarQube Python Plugin
+ * Copyright (C) 2011-2020 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.python.checks;
+
+import java.util.Arrays;
+import org.junit.Test;
+import org.sonar.python.checks.utils.PythonCheckVerifier;
+
+public class UnresolvedSymbolsCheckTest {
+
+  @Test
+  public void test() {
+    PythonCheckVerifier.verify("src/test/resources/checks/unresolvedSymbols/unresolvedSymbols.py", new UnresolvedSymbolsCheck());
+  }
+
+  @Test
+  public void test_wildcard_import() {
+    PythonCheckVerifier.verify(
+            Arrays.asList("src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsWithWildcardImport.py","src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsImported.py" ),
+            new UnresolvedSymbolsCheck());
+  }
+
+  @Test
+  public void test_wildcard_import_all_property() {
+    PythonCheckVerifier.verifyNoIssue(
+            Arrays.asList("src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsImportWithAll.py", "src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsAll/__init__.py"),
+            new UnresolvedSymbolsCheck());
+  }
+
+  @Test
+  public void test_unresolved_wildcard_import() {
+    PythonCheckVerifier.verifyNoIssue("src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsWithUnresolvedWildcardImport.py", new UnresolvedSymbolsCheck());
+  }
+
+  @Test
+  public void test_dynamic_globals() {
+    PythonCheckVerifier.verifyNoIssue("src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsWithGlobals.py", new UnresolvedSymbolsCheck());
+  }
+
+}
