--- python-checks/src/test/resources/checks/undeclaredNameUsageImportWithAll.py
@@ -1,2 +1,0 @@
-from undeclaredNameUsageAll import *
-f() # OK
