+++ python-checks/src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsAll/__init__.py
