+++ python-checks/src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsImportWithAll.py
@@ -1,0 +1,2 @@
+from unresolvedSymbolsAll import *
+f() # OK
