+++ python-checks/src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsImported.py
