+++ python-checks/src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsWithGlobals.py
