+++ python-checks/src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsWithUnresolvedWildcardImport.py
