+++ python-checks/src/test/resources/checks/unresolvedSymbols/unresolvedSymbolsWithWildcardImport.py
@@ -1,4 +1,4 @@
-from undeclaredNameUsageImported import *
+from unresolvedSymbolsImported import *
 
 def f():
     print(a) # OK, mod exports symbol a
